[Feat] Scheme+Template으로 Scheme 분리 #18
